extract `useSSRQuery`
